Account for new py3 compatibility
When pull-request #2 is merged, the download-script should be
compatible with both python 2 and python 3. Hence it is unnecessary to
explicitly favour Python 2 here.

I also implemented @k-nut 's input and added the suggested sentence,
elaborating on how to install the dependency.

I have chosen to put the second sentence on a new line. This should make
no difference in how the markdown is rendered, since it requires an
actually empty line to render a new paragraph. However it should make
future diffs more readable.